feat: select element using data-test tag
Since aria-labels can frequently change and break bot actions, this commit
keeps them as fallbacks but adds the data-test tags as the primary selectors.
It is important to note that not all the aria-labels have an equivalent
data-test tag, so some elements may still rely on aria-labels as fallback
selectors.